# add thread locks to mitigate task migration