@Moelf 2023-05-31
this one has better performance

## this PR

```julia
julia> @benchmark main(t)
BenchmarkTools.Trial: 2098 samples with 1 evaluation.
 Range (min … max):  2.149 ms …   4.569 ms  ┊ GC (min … max): 0.00% … 27.81%
 Time  (median):     2.261 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.379 ms ± 355.874 μs  ┊ GC (mean ± σ):  3.34% ±  7.78%

  ▃▇▇█▇▆▆▅▃▄▃▂▂                                  ▁ ▁  ▁       ▁
  █████████████▇▇▇█▇▅▇▆▅▅▅▁▅▅▅▁▁▁▄▁▁▁▁▁▁▁▁▄▄▅▅▅▆▇██████▇█▆▄▅▅ █
  2.15 ms      Histogram: log(frequency) by time      3.72 ms <

 Memory estimate: 3.36 MiB, allocs estimate: 4059.
```

## before this PR

```julia

julia> @benchmark main(t)
BenchmarkTools.Trial: 2474 samples with 1 evaluation.
 Range (min … max):  1.845 ms …   3.709 ms  ┊ GC (min … max): 0.00% … 31.37%
 Time  (median):     1.929 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.018 ms ± 315.950 μs  ┊ GC (mean ± σ):  3.71% ±  8.39%

  ▂▇▇█▇▅▄▂▁                                             ▁ ▂   ▁
  ██████████▇▆▁▄▁▄▄▆▄▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▅▇▇████████ █
  1.85 ms      Histogram: log(frequency) by time      3.18 ms <

 Memory estimate: 3.36 MiB, allocs estimate: 4059.
```