@Moelf 2023-06-01
I think technically `br = ...[tid]` need also be in the lock